+++ History.markdown
@@ -1,5 +1,7 @@
-## HEAD
+## 1.0.4 / 2013-07-24
+<<<<<<< HEAD
 
+<<<<<<< HEAD
 ### Minor Enhancements
 
   * Add 'sample' Liquid filter Equivalent to Array#sample functionality (#4223)
@@ -459,6 +461,10 @@
   * In the case that a Gemfile does not exist, ensure Jekyll doesn't fail on requiring the Gemfile group (#3066)
 
 ### Development Fixes
+=======
+### Bug Fixes
+  * Require Liquid 2.5.1 (#1349)
+>>>>>>> jekyll/1.0-branch
 
   * Fix a typo in the doc block for `Jekyll::URL.escape_path` (#3052)
   * Add integration test for `jekyll new --blank` in TestUnit (#2913)
@@ -583,6 +589,7 @@
 
 ### Site Enhancements
 
+<<<<<<< HEAD
   * Use Sass and a Docs Collection (#2651)
   * Add `latest_version.txt` file to the site (#2740)
   * Be more ambiguous about `page.content`. But more transparent. (#2522)
@@ -1000,11 +1007,17 @@
   * Add docs for `where` and `group_by` Liquid filters (#2298)
   * Remove notes in docs for unreleased features (#2309)
 
+=======
+>>>>>>> jekyll/v1-stable
 ## 1.5.1 / 2014-03-27
 
 ### Bug Fixes
 
+<<<<<<< HEAD
   * Only strip the drive name if it begins the string (#2176)
+=======
+* Only strip the drive name if it begins the string (#2176)
+>>>>>>> jekyll/v1-stable
 
 ## 1.5.0 / 2014-03-24
 
@@ -1020,11 +1033,19 @@
 
 ### Development Fixes
 
+<<<<<<< HEAD
   * Lock `cucumber` at `1.3.11` (#2167)
+=======
+  * Lock `cucmber` at `1.3.11` (#2167)
+>>>>>>> jekyll/v1-stable
 
 ## 1.4.3 / 2014-01-13
 
 ### Bug Fixes
+<<<<<<< HEAD
+=======
+
+>>>>>>> jekyll/v1-stable
   * Patch show-stopping security vulnerabilities (#1944)
 
 ## 1.4.2 / 2013-12-16
@@ -1357,6 +1378,11 @@
   * Add 'Working with drafts' page to docs (#1289)
   * Add information about time zones to the documentation for a page's
     date (#1304)
+=======
+
+### Bug Fixes
+  * Require Liquid 2.5.1 (#1349)
+>>>>>>> origin/1.0-branch
 
 ## 1.0.3 / 2013-06-07